refactor: :recycle: cria a pasta utils para reuso da função formatDate
